+++ src/viewer/tsdb/series/mod.rs
@@ -1,0 +1,9 @@
+use super::*;
+
+mod counter;
+mod gauge;
+mod untyped;
+
+pub use counter::CounterSeries;
+pub use gauge::GaugeSeries;
+pub use untyped::UntypedSeries;
